refactor : api 경로 변경 및 필요없는 코드 정리 (#97)
* refactor : api 경로 변경 코드 정리

* feat : 문제 풀이 성공시 캐릭터 레벨업 이벤트 publisher 추가

* bug : 웹소켓 disconnect 시 채팅 관련 에러 뜨는 문제 해결

+++ src/main/java/org/ezcode/codetest/application/chatting/service/ChattingUseCase.java
@@ -56,7 +56,7 @@ public class ChattingUseCase {
 
 		ChatRoom removedRoom = chattingDomainService.getChatRoom(request.roomId());
 
-		chattingDomainService.isChatRoomOwner(removedRoom, user.getId());
+		chattingDomainService.checkChatRoomOwnerOrAdmin(removedRoom, user);
 
 		chattingDomainService.removeChatRoom(removedRoom);
