+++ src/main/java/org/ezcode/codetest/infrastructure/event/listener/WebSocketEventListener.java
@@ -8,7 +8,7 @@ import org.ezcode.codetest.application.chatting.service.ChatMessageTemplate;
 import org.ezcode.codetest.domain.chat.model.ChatRoom;
 import org.ezcode.codetest.domain.chat.service.ChattingDomainService;
 import org.ezcode.codetest.domain.user.service.UserDomainService;
-import org.ezcode.codetest.infrastructure.event.service.StompMessageService;
+import org.ezcode.codetest.infrastructure.event.publisher.StompMessageService;
 import org.ezcode.codetest.infrastructure.session.service.RedisSessionCountService;
 import org.springframework.context.ApplicationListener;
 import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
@@ -31,11 +31,24 @@ public class WebSocketEventListener implements ApplicationListener<SessionDiscon
 	@Override
 	public void onApplicationEvent(SessionDisconnectEvent event) {
 
-		try {
-			StompHeaderAccessor h = StompHeaderAccessor.wrap(event.getMessage());
-			String sessionId = h.getSessionId();
+		StompHeaderAccessor header = StompHeaderAccessor.wrap(event.getMessage());
 
-			String email = h.getUser().getName();
+		Map<String, Object> attrs = header.getSessionAttributes();
+
+		if (attrs == null) {
+			return;
+		}
+
+		Boolean isChatWs = (Boolean)attrs.get("isChattingWebsocket");
+
+		if (!Boolean.TRUE.equals(isChatWs)) {
+			return;
+		}
+
+		try {
+			String sessionId = header.getSessionId();
+
+			String email = header.getUser().getName();
 			String nickName = userDomainService.getUser(email).getNickname();
 
 			RoomSessionInfo roomData = sessionService.removeSessionCount(sessionId);
@@ -51,7 +64,7 @@ public class WebSocketEventListener implements ApplicationListener<SessionDiscon
 			messageService.handleChatRoomEntryExitMessage(ChatMessageTemplate.CHAT_ROOM_LEFT.format(nickName),
 				chatRoom.getId());
 		} catch (Exception e) {
-			log.info("SessionDisconnectEvent 처리 중 예외 발생, 채팅 관련 웹소켓 세션이 아닙니다.", e);
+			log.info("SessionDisconnectEvent 채팅 세션 정리중 예외 발생", e);
 		}
 	}
 }
